requirements.txt
Flask==3.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
gunicorn==21.2.0
